BigInteger to Long

+++ lib/src/main/java/org/eclipse/tsp/java/client/shared/query/QueryInterval.java
@@ -1,7 +1,5 @@
 package org.eclipse.tsp.java.client.shared.query;
 
-import java.math.BigInteger;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -14,9 +12,9 @@ import lombok.RequiredArgsConstructor;
 @NoArgsConstructor
 public class QueryInterval {
 	@NonNull
-	private BigInteger start;
+	private Long start;
 	@NonNull
-	private BigInteger end;
+	private Long end;
 	private Integer nbTimes;
 
 }
